refactor: split the Crowdin normalizer into one export per file
