+++ packages/twenty-utils/crowdin-normalizer/errors/crowdin-api.error.ts
@@ -1,0 +1,8 @@
+export class CrowdinApiError extends Error {
+  constructor(
+    readonly status: number,
+    readonly body: string,
+  ) {
+    super(`Crowdin API error: ${status} ${body}`);
+  }
+}
