+++ packages/twenty-utils/crowdin-normalizer/types/crowdin-context.type.ts
@@ -1,0 +1,4 @@
+export type CrowdinContext = {
+  token: string;
+  projectId: number;
+};
