+++ packages/twenty-utils/crowdin-normalizer/types/crowdin-translation.type.ts
@@ -1,0 +1,5 @@
+export type CrowdinTranslation = {
+  stringId: number;
+  translationId: number;
+  text: string;
+};
