+++ packages/twenty-utils/crowdin-normalizer/types/normalization-rule.type.ts
@@ -1,0 +1,6 @@
+export type NormalizationRule = {
+  name: string;
+  detect: (text: string) => boolean;
+  fix: (text: string) => string;
+  sourceFilter?: (sourceText: string) => boolean;
+};
